--- src/main/java/com/example/hhproject0/Hhproject0Application.java
@@ -1,13 +1,0 @@
-package com.example.hhproject0;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Hhproject0Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Hhproject0Application.class, args);
-    }
-
-}
