--- src/main/java/com/example/hhproject0/user/controller/UserController.java
@@ -1,4 +1,0 @@
-package com.example.hhproject0.user.controller;
-
-public class UserController {
-}
