--- src/main/java/com/example/hhproject0/user/dto/UserDto.java
@@ -1,4 +1,0 @@
-package com.example.hhproject0.user.dto;
-
-public class UserDto {
-}
