--- src/main/java/com/example/hhproject0/user/entity/User.java
@@ -1,25 +1,0 @@
-package com.example.hhproject0.user.entity;
-
-import lombok.Getter;
-
-import javax.persistence.*;
-
-@Entity(name = "users")
-@Getter
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String username;
-
-    @Column(nullable = false)
-    @Enumerated(value = EnumType.STRING)
-    private UserRoleEnum role;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-}
