--- src/main/java/com/example/hhproject0/user/exception/UserException.java
@@ -1,11 +1,0 @@
-package com.example.hhproject0.user.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UserException extends RuntimeException {
-    private final UserErrorCode errorCode;
-
-}
