--- src/main/java/com/example/hhproject0/user/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.example.hhproject0.user.repository;
-
-import com.example.hhproject0.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User,Long> {
-
-    Optional<User> findByEmail(String email);
-}
