--- src/main/java/com/example/hhproject0/user/service/UserService.java
@@ -1,4 +1,0 @@
-package com.example.hhproject0.user.service;
-
-public class UserService {
-}
