+++ src/main/java/com/example/lablink/LabLinkApplication.java
@@ -1,0 +1,13 @@
+package com.example.lablink;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LabLinkApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(LabLinkApplication.class, args);
+    }
+
+}
