+++ src/main/java/com/example/lablink/security/SecurityExceptionDto.java
@@ -1,0 +1,17 @@
+package com.example.lablink.security;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class SecurityExceptionDto {
+
+    private int statusCode;
+    private String msg;
+
+    public SecurityExceptionDto(int statusCode, String msg) {
+        this.statusCode = statusCode;
+        this.msg = msg;
+    }
+}
