+++ src/main/java/com/example/lablink/user/controller/UserController.java
@@ -1,0 +1,4 @@
+package com.example.lablink.user.controller;
+
+public class UserController {
+}
