+++ src/main/java/com/example/lablink/user/dto/UserDto.java
@@ -1,0 +1,4 @@
+package com.example.lablink.user.dto;
+
+public class UserDto {
+}
