+++ src/main/java/com/example/lablink/user/entity/User.java
@@ -1,0 +1,25 @@
+package com.example.lablink.user.entity;
+
+import lombok.Getter;
+
+import javax.persistence.*;
+
+@Entity(name = "users")
+@Getter
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, unique = true)
+    private String username;
+
+    @Column(nullable = false)
+    @Enumerated(value = EnumType.STRING)
+    private UserRoleEnum role;
+
+    @Column(nullable = false, unique = true)
+    private String email;
+
+}
