+++ src/main/java/com/example/lablink/user/entity/UserEntity.java
@@ -1,0 +1,4 @@
+package com.example.lablink.user.entity;
+
+public class UserEntity {
+}
