+++ src/main/java/com/example/lablink/user/exception/UserExceptionHandler.java
@@ -1,0 +1,19 @@
+package com.example.lablink.user.exception;
+
+
+import com.example.lablink.message.ResponseMessage;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@Slf4j
+@RestControllerAdvice
+public class UserExceptionHandler {
+    @ExceptionHandler(value = { UserException.class })
+    protected ResponseEntity<ResponseMessage> handleCustomException(UserException e) {
+        log.error("handleCustomException throw CustomException : {}", e.getErrorCode());
+        return ResponseMessage.ErrorResponse(e.getErrorCode());
+
+    }
+}
