+++ src/main/java/com/example/lablink/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.example.lablink.user.repository;
+
+import com.example.lablink.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User,Long> {
+    Optional<User> findByEmail(String email);
+}
