+++ src/main/java/com/example/lablink/user/service/UserService.java
@@ -1,0 +1,4 @@
+package com.example.lablink.user.service;
+
+public class UserService {
+}
