--- src/test/java/com/example/hhproject0/Hhproject0ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.hhproject0;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Hhproject0ApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
