+++ src/test/java/com/example/lablink/LabLinkApplication.java
@@ -1,0 +1,13 @@
+package com.example.lablink;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LabLinkApplication {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
